pyproject.toml
[tool.poetry]
name = "flaticon"
version = "0.1.0"
description = "Download de icones do site flaticon."
authors = ["matheuslpolidoro <mattpolidoro4@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.4.0"
webdriver-manager = "^3.8.3"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
mkdocs = "^1.3.1"
blue = "^0.9.1"
prospector = "^1.7.7"
isort = "^5.10.1"
pip-audit = "^2.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
